more

+++ ddh/ble_dox.py
@@ -18,7 +18,7 @@ from utils.ddh_common import (
     ddh_does_do_not_rerun_file_flag_exist, EV_BLE_DL_ERROR,
     TESTMODE_FILENAME_PREFIX,
     calculate_path_to_folder_within_dl_files_from_mac_address,
-    ddh_config_does_flag_file_download_test_mode_exist, exp_get_conf_dox,
+    ddh_config_does_flag_file_download_test_mode_exist, exp_get_conf_dox, exp_debug_skip_hbw,
 )
 from ddh_log import lg_ble as lg
 
@@ -100,23 +100,24 @@ async def ble_download_dox(d):
 
     # feature has-logger-been-in-water
     flag_ignore_hbw = ddh_get_template_of_path_of_hbw_flag_file().format(mac)
-    if state == 'running':
-        if v >= MIN_VERSION_HBW_CMD:
-            lg.a('sending command Has-Been-in-Water')
-            rv, v = await cmd_hbw()
-            _rae(rv, "hbw")
-            lg.a(f"HBW | {v}")
-            if os.path.exists(flag_ignore_hbw):
-                os.unlink(flag_ignore_hbw)
-                lg.a('file flag to ignore HBW exists, force download it')
-            else:
-                if v == 0:
-                    lg.a('logger has NOT been in water, no need to download it')
-                    await disconnect()
-                    return 2
-                lg.a("logger has been in water, we will download it")
-    else:
-        lg.a('logger NOT running, not considering HBW command')
+    if exp_debug_skip_hbw() != 1:
+        if state == 'running':
+            if v >= MIN_VERSION_HBW_CMD:
+                lg.a('sending command Has-Been-in-Water')
+                rv, v = await cmd_hbw()
+                _rae(rv, "hbw")
+                lg.a(f"HBW | {v}")
+                if os.path.exists(flag_ignore_hbw):
+                    os.unlink(flag_ignore_hbw)
+                    lg.a('file flag to ignore HBW exists, force download it')
+                else:
+                    if v == 0:
+                        lg.a('logger has NOT been in water, no need to download it')
+                        await disconnect()
+                        return 2
+                    lg.a("logger has been in water, we will download it")
+        else:
+            lg.a('logger NOT running, not considering HBW command')
 
 
     # to know if this DO-X logger uses LID or LIX files
